Deleted init_data_helper.c and moved it's content into init_data.c
