+++ parser/init_data.c
@@ -6,7 +6,7 @@
 /*   By: jjesberg <jjesberg@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/07 02:26:29 by jjesberg          #+#    #+#             */
-/*   Updated: 2023/01/17 17:21:51 by nlouro           ###   ########.fr       */
+/*   Updated: 2023/01/17 17:24:19 by nlouro           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -82,6 +82,57 @@ static void	save_param(t_data *data, char **raw_data)
 	if (data->c_colour == 0 || data->f_colour == 0)
 		exit(cub_error(COLOUR_ERROR));
 	find_map(data, raw_data);
+}
+
+static int	count_lines(char *path)
+{
+	int		count;
+	int		fd;
+	char	*line;
+
+	fd = open(path, O_RDONLY);
+	count = 1;
+	line = get_next_line(fd, 1);
+	if (!line)
+		exit(cub_error(FILE_ERROR));
+	while (line)
+	{
+		free(line);
+		count++;
+		line = get_next_line(fd, 0);
+	}
+	close(fd);
+	return (count);
+}
+
+/*
+ * read the user filename and return an array of strings representing each line of the file
+ */
+static char	**get_map(char *path)
+{
+	char	**map;
+	char	*line;
+	int		i;
+	int		fd;
+	int		len;
+
+	map = malloc(sizeof(char *) * (count_lines(path)));
+	if (!map)
+		exit(cub_error(MALLOC_ERROR));
+	i = 0;
+	fd = open(path, O_RDONLY);
+	line = get_next_line(fd, 1);
+	while (line)
+	{
+		len = ft_strlen(line) - 1;
+		if (line[len] == '\n')
+			line[len] = '\0';
+		map[i++] = line;
+		line = get_next_line(fd, 0);
+	}
+	map[i] = NULL;
+	close(fd);
+	return (map);
 }
 
 /*
